monev selesai yaallah akhirnya

+++ database/migrations/2025_08_27_005225_create_jawaban_users_table.php
@@ -16,8 +16,8 @@ return new class extends Migration
             $table->foreignId('opsi_jawaban_id')->nullable()->constrained('opsi_jawabans')->onDelete('cascade');
             $table->foreignId('pertanyaan_id')->constrained('pertanyaans')->onDelete('cascade');
             $table->foreignId('percobaan_id')->constrained('percobaans')->onDelete('cascade');
-            $table->tinyInteger('nilai_jawaban')->nullable(); // Untuk skala likert (1-5)
-                $table->text('jawaban_teks')->nullable(); // Untuk jawaban esai / teks bebas
+            $table->integer('nilai_jawaban')->nullable(); // Untuk skala likert (1-5)
+            $table->text('jawaban_teks')->nullable(); // Untuk jawaban esai / teks bebas
             $table->timestamps();
         });
     }
